requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
xgboost>=2.0
imbalanced-learn>=0.11
matplotlib>=3.7
seaborn>=0.12
joblib>=1.3
pyyaml>=6.0
pytest>=7.0
pydantic>=2.0
ruff>=0.4
